add new logo and set custom colors
