+++ app/src/main/java/com/kodeco/android/coordplot/MainActivity.kt
@@ -1,5 +1,6 @@
 package com.kodeco.android.coordplot
 
+import android.content.res.Resources.Theme
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
@@ -77,7 +78,7 @@ fun Map(xPercent: Float, yPercent: Float, modifier: Modifier = Modifier) {
     Box(
         modifier = modifier
             .size(300.dp)
-            .background(MaterialTheme.colorScheme.secondary),
+            .background(MaterialTheme.colorScheme.tertiary),
     ) {
         // TODO Swap this Box out for an Image.
         //  Make sure to keep the size and offset modifiers.
@@ -86,10 +87,10 @@ fun Map(xPercent: Float, yPercent: Float, modifier: Modifier = Modifier) {
                 .size(36.dp)
                 .offset((xPercent * 300 - 18).dp, (yPercent * 300 - 18).dp)
                 .clip(CircleShape)
-                .background(Color.Green)
+                .background(Color.Blue)
         )
         Image(
-            painter = painterResource(id = R.drawable.star),
+            painter = painterResource(id = R.drawable.logo),
             contentDescription = "Star pin",
             modifier = Modifier
                 .size(36.dp)
@@ -106,25 +107,3 @@ fun MapPreview() {
     }
 }
 
-// TODO Move MapSlider and MapSliderPreview() to their own file
-@Composable
-fun MapSlider(titleRes: Int, percentage: Float, onPercentageChange: (Float) -> Unit) {
-    // TODO Provide a Row composable with an inner Text composable for the title
-    //  and move the Slider composable used in PlotSurface here.
-    //  - The Row should have a horizontal padding of 16.dp
-    //  - The Text title should have a start padding of 8.dp with a width
-    //    of 120.dp.
-    //  - The Slider should wire up its onValueChange with the MapSlider's
-    //    onPercentageChange lambda.
-}
-
-@Preview(showBackground = true)
-@Composable
-fun MapSliderPreview() {
-    MyApplicationTheme {
-        MapSlider(
-            titleRes = R.string.app_name, // TODO replace this with a valid string resource
-            percentage = 0.5f,
-        ) {}
-    }
-}
