+++ app/src/main/java/com/kodeco/android/coordplot/ui/theme/Color.kt
@@ -2,11 +2,6 @@ package com.kodeco.android.coordplot.ui.theme
 
 import androidx.compose.ui.graphics.Color
 
-// TODO make some custom colors
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
+val Tomato = Color(0xFFFF6347)
+val DarkBlue = Color(0xFF003366)
+val ForestGreen = Color(0xFF568203)
