+++ app/src/main/res/drawable/logo.xml
@@ -1,0 +1,18 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="72dp"
+    android:height="72dp"
+    android:viewportWidth="72"
+    android:viewportHeight="72">
+  <path
+      android:pathData="M36,0L36,0A36,36 0,0 1,72 36L72,36A36,36 0,0 1,36 72L36,72A36,36 0,0 1,0 36L0,36A36,36 0,0 1,36 0z"
+      android:strokeWidth="1"
+      android:fillColor="#F7CD34"
+      android:fillType="evenOdd"
+      android:strokeColor="#00000000"/>
+  <path
+      android:pathData="M38.878,33.12C33.292,31.44 32.416,30.327 32.416,28.691C32.416,27.251 33.708,25.985 36.271,25.985C39.557,25.985 40.762,27.84 41.988,30.393L48.91,26.356C46.304,21.076 42.164,18 36.271,18C33.204,18 30.203,18.982 27.969,20.815L15.855,20.815L15.855,22.691L26.173,22.691C25.537,23.564 25.034,24.524 24.705,25.549L12,25.549L12,27.425L24.289,27.425C24.223,27.905 24.201,28.385 24.201,28.887C24.201,29.345 24.223,29.825 24.289,30.284L15.855,30.284L15.855,32.182L24.683,32.182C26.085,36.611 30.4,38.553 34.781,39.775C40.17,41.258 42.229,42.131 42.229,44.313C42.229,45.687 41.265,47.215 37.344,47.215C33.292,47.215 31.452,45.076 30.159,42.109L23.062,46.211C25.056,51.469 29.524,55.2 37.06,55.2C44.858,55.2 50.4,51.12 50.4,44.204C50.422,36.655 44.31,34.756 38.878,33.12Z"
+      android:strokeWidth="1"
+      android:fillColor="#FFFFFF"
+      android:fillType="nonZero"
+      android:strokeColor="#00000000"/>
+</vector>
